requirements.txt
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tldextract==5.1.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
